% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{collapse}
\alias{collapse}
\title{Collapses a SSN object into a data frame}
\usage{
collapse(t)
}
\arguments{
\item{t}{Path to a SSN object}
}
\value{
A data frame
}
\description{
Collapses a SSN object into a data frame
}
\examples{
#t.df <- collapse(t)

}
