% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{convert_network_to_graph}
\alias{convert_network_to_graph}
\title{Combines two `sf` objects, the sensor locations and river network, in to an `sfnetwork`.}
\usage{
convert_network_to_graph(river_network, sensor_locations)
}
\arguments{
\item{river_network}{An `sf` object of `MULTILINESTRING`(s) containing the streams and rivers, cropped to the sensor locations.}

\item{sensor_locations}{An `sf` dataframe containing sensor locations and `locID`.}
}
\value{
An `sfnetwork` graph containing the river network and sensors.
}
\description{
Combines two `sf` objects, the sensor locations and river network, in to an `sfnetwork`.
}
\details{
This function "snaps" sensor locations to the nearest point on the network.
Additionally, the sensor locations are stored in the graph, `g`, and can be accessed via `V(g)$type`, as "sensor".
}
\author{
Sean Francis
}
