% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{dist_wei_mat}
\alias{dist_wei_mat}
\title{Creates a list of distances and weights}
\usage{
dist_wei_mat(path = path)
}
\arguments{
\item{path}{Path to the files}
}
\value{
A list of matrices
}
\description{
Creates a list of distances and weights
}
\examples{
#mat_all <- dist_wei_mat(path)

}
