% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{format_sensor_data}
\alias{format_sensor_data}
\title{Converts a dataframe containing sensor data in to an `sf` object}
\usage{
format_sensor_data(sensor_data, lon_name, lat_name)
}
\arguments{
\item{sensor_data}{A dataframe containing sensor locations and a column called `locID` denoting the location IDs.}

\item{lon_name}{The column name containing the longitude of the sensor.}

\item{lat_name}{The column name containing the latitude of the sensor.}
}
\value{
An `sf` dataframe containing sensor locations and `locID`.
}
\description{
Converts a dataframe containing sensor data in to an `sf` object
}
\author{
Sean Francis
}
