% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{gen_distance_matrices}
\alias{gen_distance_matrices}
\title{Creates a list containing the stream distances and weights}
\usage{
gen_distance_matrices(network, sensor_locations, root_loc)
}
\arguments{
\item{network}{An `sfnetwork` containing the river network and sensor locations.}

\item{sensor_locations}{An `sf` dataframe containing sensor locations and `locID`.}

\item{root_loc}{A sensor location which all water flows towards.}
}
\value{
A list of matrices.
}
\description{
Creates a list containing the stream distances and weights
}
\author{
Sean Francis
}
