% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{generate_osm_ssn}
\alias{generate_osm_ssn}
\title{Generates an Open Street Maps Spatial Stream Network.}
\usage{
generate_osm_ssn(
  sensor_data,
  lon_name,
  lat_name,
  root_loc,
  plot_network = TRUE,
  gen_pred_sites = FALSE,
  num_pred_sites = NA
)
}
\arguments{
\item{sensor_data}{A dataframe containing sensor locations and a column called `locID` denoting the location IDs.}

\item{lon_name}{The column name containing the longitude of the sensor.}

\item{lat_name}{The column name containing the latitude of the sensor.}

\item{root_loc}{A sensor location which all water flows towards.}

\item{plot_network}{A bool indicating whether to plot the network. Useful for determining potential errors. Will be saved in object.}

\item{gen_pred_sites}{A bool indicating whether to generate new prediction sites.}

\item{num_pred_sites}{Required if `gen_pred_sites=TRUE`. An integer indicating the number of prediction sites to generate.}
}
\value{
An `osm_ssn` object, containing: the graph network as an `sfnetwork`; the distance adjacency matrices; if `plot_network=TRUE`, an interactive leaflet plot, if `gen_pred_sites=TRUE`, an `sf` object containing the locations of the predictions sites.
}
\description{
It will take a dataframe containing sensor locations, with a coulmn `locID` denoting location IDs (can be string or numeric).
Requires a root sensor location, in which all sensor locations are connected to, and flow towards.
}
\details{
Assumes longitude and latitude data are in the 4326 coordinate reference system. If they are not, data will need to be converted beforehand using `st_transform()`. Predictions on sites using gen_pred_sites is under development specially for tail-down models.
}
\examples{
\donttest{
# require('SSNdata')
# clear <- readRDS(system.file("extdata/clear_obs.RDS", package = "SSNdata"))
#
# # Generate osm ssn
# clear_osm_ssn <- generate_osm_ssn(clear,
#                                   "long", "lat",
#                                   root_loc = 12,
#                                   plot_network = TRUE,
#                                   gen_pred_sites = TRUE,
#                                   num_pred_sites = 20)
# # Show plot
# clear_osm_ssn$plot
}
}
\author{
Sean Francis
}
