% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{generate_prediction_locations}
\alias{generate_prediction_locations}
\title{Creates an `sf` dataframe containing equally spaced points on a river network}
\usage{
generate_prediction_locations(river_network, num_pred_sites)
}
\arguments{
\item{river_network}{An `sf` object of `MULTILINESTRING`(s) containing the streams and rivers, cropped to the sensor locations.}

\item{num_pred_sites}{An integer indicating the number of prediction sites to generate.}
}
\value{
An `sf` dataframe of `POINT`(s) containing prediction sites.
}
\description{
Creates an `sf` dataframe containing equally spaced points on a river network
}
\author{
Sean Francis
}
