% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{generate_river_network}
\alias{generate_river_network}
\title{Takes an `sf` object as input, and queries Open Street Maps for rivers and streams within the bounds of the sensor locations.}
\usage{
generate_river_network(sensor_locations, root_loc)
}
\arguments{
\item{sensor_locations}{An `sf` dataframe containing sensor locations and `locID`.}

\item{root_loc}{A sensor location which all water flows towards.}
}
\value{
An `sf` object of `MULTILINESTRING`(s) containing the streams and rivers, cropped to the sensor location, with all lines not containing the outlet/root_loc removed.
}
\description{
Takes an `sf` object as input, and queries Open Street Maps for rivers and streams within the bounds of the sensor locations.
}
\details{
This function queries Open Street Maps using `osmdata` package in R. Crops the data to the bounding box of the sensor locations. Removes all lines not connected to the outlet/root_loc.
}
\author{
Sean Francis
}
