% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{mylm}
\alias{mylm}
\title{A simple modeling function using a formula and data}
\usage{
mylm(formula, data)
}
\arguments{
\item{formula}{A formula as in lm()}

\item{data}{A data.frame containing the elements specified in the formula}
}
\value{
A list of matrices
}
\description{
A simple modeling function using a formula and data
}
\examples{
#options(na.action='na.pass')
#out_list = mylm(formula = y ~ X1 + X2 + X3, data = data)

}
\author{
Jay ver Hoef
}
