% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{ssnbayes}
\alias{ssnbayes}
\title{Fits the model using Stan}
\usage{
ssnbayes(
  formula = formula,
  data = data,
  CorModels = "Exponential.tailup",
  iter = 3000,
  warmup = 1500,
  chains = 3,
  refresh = max(iter/100, 1)
)
}
\arguments{
\item{formula}{A formula as in lm()}

\item{data}{A list}

\item{CorModels}{Correlation structure}

\item{iter}{Number of iterations}

\item{warmup}{Warm up samples}

\item{chains}{Number of chains}

\item{refresh}{Sampler refreshing rate}
}
\value{
A list with the fit
}
\description{
Fits the model using Stan
}
\examples{
#fit_td <- ssnbayes(formula = 'y ~ X1 + X2 + X3',
#                     data = data,
#                    CorModels = "Exponential.taildown",
#                     iter = 3000,
#                     warmup = 1500,
#                    chains = 3)
}
