% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{collapse}
\alias{collapse}
\title{Collapses a SSN object into a data frame}
\usage{
collapse(t, par = "afvArea")
}
\arguments{
\item{t}{Path to a SSN object}

\item{par}{A spatial parameter}
}
\value{
A data frame
}
\description{
Collapses a SSN object into a data frame
}
\examples{
\donttest{
require("SSN")
path <- system.file("extdata/clearwater.ssn", package = "SSNbayes")
n <- importSSN(path, predpts = "preds", o.write = TRUE)
t.df <- collapse(n)}
}
