% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{dist_wei_mat}
\alias{dist_wei_mat}
\title{Creates a list of distances and weights}
\usage{
dist_wei_mat(path = path, net = 1, addfunccol = "addfunccol")
}
\arguments{
\item{path}{Path to the files}

\item{net}{(optional) A network from the SSN object}

\item{addfunccol}{(optional) A parameter to compute the spatial weights}
}
\value{
A list of matrices
}
\description{
Creates a list of distances and weights
}
\examples{
\donttest{
path <- system.file("extdata/clearwater.ssn", package = "SSNbayes")
mat_all <- dist_wei_mat(path, net = 2, addfunccol='afvArea')
}
}
