% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{mylm}
\alias{mylm}
\title{A simple modeling function using a formula and data}
\usage{
mylm(formula, data)

mylm(formula, data)
}
\arguments{
\item{formula}{A formula as in lm()}

\item{data}{A data.frame containing the elements specified in the formula}
}
\value{
A list of matrices

A list of matrices
}
\description{
A simple modeling function using a formula and data

A simple modeling function using a formula and data
}
\examples{
options(na.action='na.pass')
data("iris")
out_list = mylm(formula = Petal.Length ~ Sepal.Length + Sepal.Width, data = iris)
options(na.action='na.pass')
data("iris")
out_list = mylm(formula = Petal.Length ~ Sepal.Length + Sepal.Width, data = iris)
}
\author{
Jay ver Hoef
}
