% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_func.R
\name{ssnbayes}
\alias{ssnbayes}
\title{Fits a mixed regression model using Stan
It requires the same number of observation/locations per day.
Missing values are alloed in the response but not in the covariates.}
\usage{
ssnbayes(
  formula = formula,
  data = data,
  ssn = ssn,
  CorModels = "Exponential.tailup",
  iter = 3000,
  warmup = 1500,
  chains = 3,
  refresh = max(iter/100, 1)
)
}
\arguments{
\item{formula}{A formula as in lm()}

\item{data}{A data.frame containing the locations, dates, covariates and the response variable}

\item{ssn}{A .SSN object}

\item{CorModels}{Correlation structure}

\item{iter}{Number of iterations}

\item{warmup}{Warm up samples}

\item{chains}{Number of chains}

\item{refresh}{Sampler refreshing rate}
}
\value{
A list with the fit
}
\description{
Fits a mixed regression model using Stan
It requires the same number of observation/locations per day.
Missing values are alloed in the response but not in the covariates.
}
\examples{
#fit_td <- ssnbayes(formula = 'y ~ X1 + X2 + X3',
#                    data = data, ssn = ssn,
#                    CorModels = "Exponential.taildown",
#                    iter = 3000,
#                    warmup = 1500,
#                    chains = 3)
}
